+++ frontend/src/components/common/TripInformation/TripInformation.tsx
@@ -1,6 +1,6 @@
 import DefaultThumbnail from '@assets/png/trip-information_default-thumbnail.png';
 import type { TripData } from '@type/trip';
-import { Badge, Box, Button, Flex, Heading, Text, Theme, useOverlay } from 'hang-log-design-system';
+import { Badge, Box, Flex, Heading, Text, Theme, useOverlay } from 'hang-log-design-system';
 import { memo } from 'react';
 
 import { formatDate } from '@utils/formatter';
@@ -8,16 +8,20 @@ import { formatDate } from '@utils/formatter';
 import {
   buttonContainerStyling,
   descriptionStyling,
-  editButtonStyling,
   imageWrapperStyling,
   sectionStyling,
   titleStyling,
 } from '@components/common/TripInformation/TripInformation.style';
 import TripInfoEditModal from '@components/trip/TripInfoEditModal/TripInfoEditModal';
 
-type TripInformationProps = Omit<TripData, 'dayLogs'>;
+import TripButtons from './TripButtons/TripButtons';
+import TripEditButtons from './TripEditButtons/TripEditButtons';
 
-const TripInformation = ({ ...information }: TripInformationProps) => {
+interface TripInformationProps extends Omit<TripData, 'dayLogs'> {
+  isEditable?: boolean;
+}
+
+const TripInformation = ({ isEditable = true, ...information }: TripInformationProps) => {
   const { isOpen: isEditModalOpen, close: closeEditModal, open: openEditModal } = useOverlay();
 
   return (
@@ -44,14 +48,11 @@ const TripInformation = ({ ...information }: TripInformationProps) => {
           </Text>
         </Box>
         <Box css={buttonContainerStyling}>
-          {/* 수정 모드일 때만 보인다 */}
-          <Button onClick={openEditModal} css={editButtonStyling} variant="outline" size="small">
-            여행 정보 수정
-          </Button>
-          {/* TODO : 클릭하면 읽기 전용 페이지로 이동 i.e. /trip/1 */}
-          <Button variant="primary" size="small">
-            저장
-          </Button>
+          {isEditable ? (
+            <TripEditButtons tripId={information.id} openEditModal={openEditModal} />
+          ) : (
+            <TripButtons tripId={information.id} />
+          )}
         </Box>
       </header>
       {isEditModalOpen && (
