+++ frontend/src/constants/path.ts
@@ -2,7 +2,7 @@ export const PATH = {
   ROOT: '/',
   CREATE_TRIP: '/trip-new',
   EDIT_TRIP: '/trip-edit/:tripId',
-  TRIP: '/trip',
+  TRIP: '/trip/:tripId',
   EXPENSE: '/trip/expense',
   RELOAD: 0,
 } as const;
